Update
modifications
-3 vulnerability removed
-better model and controller in public
