+++ package.json
@@ -26,7 +26,7 @@
     "dotenv": "^10.0.0",
     "express": "^4.17.1",
     "method-override": "^3.0.0",
-    "nedb": "^1.8.0"
+    "nedb-promises": "^6.2.3"
   },
   "repository": {
     "type": "git",
